Adding data to the database

+++ src/main/java/com/fdryt/nursery/domain/Family.java
@@ -1,6 +1,11 @@
 package com.fdryt.nursery.domain;
 
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.SequenceGenerator;
+import jakarta.persistence.Table;
 import lombok.NoArgsConstructor;
 
 import static jakarta.persistence.GenerationType.SEQUENCE;
@@ -13,7 +18,7 @@ public class Family {
     @Id
     @GeneratedValue(strategy = SEQUENCE, generator = "family_generator")
     @SequenceGenerator(name = "family_generator")
-    private Long id;
+    private Integer id;
 
     @Column(length = 50, nullable = false, unique = true)
     private String commonName;
